third_party/blink/web_tests/fast/dom/shadow/shadowroot-type-expected.txt

CONSOLE WARNING: line 15: Element.createShadowRoot is deprecated and will be removed in M80, around February 2020. Please use Element.attachShadow instead. See https://www.chromestatus.com/features/4507242028072960 and https://developers.google.com/web/updates/2019/07/web-components-time-to-upgrade for more details.
This test ensures ShadowRootType is correctly reflected

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS "V0ShadowRoot" is internals.shadowRootType(shadowRootForDiv)
PASS "UserAgentShadowRoot" is internals.shadowRootType(shadowRootForInput)
PASS successfullyParsed is true

TEST COMPLETE
